Add tests for entities, update main template, add checker for admin roles for backend access

+++ src/Controller/PageBuilderBlockController.php
@@ -6,18 +6,25 @@ use JulianKoster\PageBuilderBundle\Entity\PageBuilderBlock;
 use JulianKoster\PageBuilderBundle\Form\PageBuilderBlockType;
 use JulianKoster\PageBuilderBundle\Repository\PageBuilderBlockRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use JulianKoster\PageBuilderBundle\Service\AdminRolesChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/block')]
 final class PageBuilderBlockController extends AbstractController
 {
-    #[Route('/index', name: 'app_admin_page_builder_block_index', methods: ['GET'])]
-    public function render_index(PageBuilderBlockRepository $repository): Response
+    #[Route('/index', name: 'juliankoster_pagebuilder_mainbuilder_render_block_index', methods: ['GET'])]
+    public function render_block_index(PageBuilderBlockRepository $repository, AdminRolesChecker $adminRolesChecker): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         return $this->render('@PageBuilderBundle/ui/block/index.html.twig', [
             'blocks' => $repository->findBy([], ["name" => "ASC"]),
         ]);
@@ -28,19 +35,25 @@ final class PageBuilderBlockController extends AbstractController
      */
     #[Route(
         path: '/crud/{operation}/{block}',
-        name: 'app_admin_page_builder_block_crud',
+        name: 'juliankoster_pagebuilder_mainbuilder_render_block_crud',
         requirements: [
             '_locale' => '%app.supported_locales%',
         ],
     )]
-    public function render_crud(
-        string $operation,
+    public function render_block_crud(
+        string                 $operation,
         EntityManagerInterface $entityManager,
-        TranslatorInterface $translator,
-        Request $request,
-        PageBuilderBlock $block = null
+        TranslatorInterface    $translator,
+        Request                $request,
+        AdminRolesChecker      $adminRolesChecker,
+        PageBuilderBlock       $block = null,
     ): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         if ($operation == "delete") {
             $entityManager->remove($block);
             $entityManager->flush();
