+++ src/Controller/PageBuilderCategoryController.php
@@ -6,18 +6,25 @@ use Doctrine\ORM\EntityManagerInterface;
 use JulianKoster\PageBuilderBundle\Entity\PageBuilderBlockCategory;
 use JulianKoster\PageBuilderBundle\Form\PageBuilderBlockCategoryType;
 use JulianKoster\PageBuilderBundle\Repository\PageBuilderBlockCategoryRepository;
+use JulianKoster\PageBuilderBundle\Service\AdminRolesChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/category')]
 class PageBuilderCategoryController extends AbstractController
 {
-    #[Route('/index', name: 'app_admin_page_builder_block_category_index', methods: ['GET'])]
-    public function render_category_index(PageBuilderBlockCategoryRepository $repository): Response
+    #[Route('/index', name: 'juliankoster_pagebuilder_mainbuilder_render_block_category_index', methods: ['GET'])]
+    public function render_category_index(PageBuilderBlockCategoryRepository $repository, AdminRolesChecker $adminRolesChecker): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         return $this->render('@PageBuilderBundle/ui/category/index.html.twig', [
             'categories' => $repository->findBy([], ["name" => "ASC"]),
         ]);
@@ -28,16 +35,22 @@ class PageBuilderCategoryController extends AbstractController
      */
     #[Route(
         path: '/crud/{operation}/{category}',
-        name: 'app_admin_page_builder_category_crud',
+        name: 'juliankoster_pagebuilder_mainbuilder_render_block_category_crud',
     )]
     public function render_crud(
-        string $operation,
-        EntityManagerInterface $entityManager,
-        TranslatorInterface $translator,
-        Request $request,
+        string                   $operation,
+        EntityManagerInterface   $entityManager,
+        TranslatorInterface      $translator,
+        Request                  $request,
+        AdminRolesChecker        $adminRolesChecker,
         PageBuilderBlockCategory $category = null
     ): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         if ($operation == "delete") {
             $entityManager->remove($category);
             $entityManager->flush();
