+++ src/Controller/PageBuilderPageController.php
@@ -6,18 +6,25 @@ use JulianKoster\PageBuilderBundle\Entity\PageBuilderPage;
 use JulianKoster\PageBuilderBundle\Form\PageBuilderPageType;
 use JulianKoster\PageBuilderBundle\Repository\PageBuilderPageRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use JulianKoster\PageBuilderBundle\Service\AdminRolesChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[Route('/page')]
 final class PageBuilderPageController extends AbstractController
 {
-    #[Route('/index', name: 'app_admin_page_builder_page_index', methods: ['GET'])]
-    public function render_index(PageBuilderPageRepository $repository): Response
+    #[Route('/index', name: 'juliankoster_pagebuilder_mainbuilder_render_page_index', methods: ['GET'])]
+    public function render_index(PageBuilderPageRepository $repository, AdminRolesChecker $adminRolesChecker): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         return $this->render('@PageBuilderBundle/ui/page/index.html.twig', [
             'pages' => $repository->findAll(),
         ]);
@@ -28,24 +35,30 @@ final class PageBuilderPageController extends AbstractController
      */
     #[Route(
         path: '/crud/{operation}/{page}',
-        name: 'app_admin_page_builder_page_crud',
+        name: 'juliankoster_pagebuilder_mainbuilder_render_page_crud',
         requirements: [
             '_locale' => '%app.supported_locales%',
         ],
     )]
     public function render_crud(
-        string $operation,
+        string                 $operation,
         EntityManagerInterface $entityManager,
-        TranslatorInterface $translator,
-        Request $request,
-        PageBuilderPage $page = null
+        TranslatorInterface    $translator,
+        Request                $request,
+        AdminRolesChecker      $adminRolesChecker,
+        PageBuilderPage        $page = null
     ): Response
     {
+        if(!$adminRolesChecker->checkRoles())
+        {
+            throw new AccessDeniedHttpException();
+        }
+
         if ($operation == "delete") {
             $entityManager->remove($page);
             $entityManager->flush();
             $this->addFlash('success', $translator->trans('Page deleted', domain: 'admin_page_builder'));
-            return $this->redirectToRoute('app_admin_page_builder_page_index');
+            return $this->redirectToRoute('juliankoster_pagebuilder_mainbuilder_render_page_index');
         }
 
         $form = $this->createForm(PageBuilderPageType::class, $page);
@@ -59,7 +72,7 @@ final class PageBuilderPageController extends AbstractController
 
                 $entityManager->flush();
                 $this->addFlash('success', $translator->trans('Page saved'));
-                return $this->redirectToRoute('app_admin_page_builder_page_index');
+                return $this->redirectToRoute('juliankoster_pagebuilder_mainbuilder_render_page_index');
             }
         }
